@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using EngrCADOCWrapper;
 
 namespace EngrCAD.Core.Nodes
@@ -10,6 +11,7 @@ namespace EngrCAD.Core.Nodes
         public abstract NativeWrapper Generate();
 
         public float CalculateVolume() => Wrapper.CalculateVolume();
-
+        public IEnumerable<Face> Faces => Wrapper.GetFaces().Select(wrapper => new Face(wrapper));
+        public IEnumerable<Edge> Edges => Wrapper.GetEdges().Select(wrapper => new Edge(wrapper));
     }
 }
