@@ -4,11 +4,13 @@
 #include <STEPControl_Reader.hxx>
 #include <STEPControl_Writer.hxx>
 #include <BRepPrimAPI_MakeSphere.hxx>
+#include <BRepOffsetAPI_MakeThickSolid.hxx>
 #include <BRepAlgoAPI_Cut.hxx>
 #include <BRepGProp.hxx>
 #include <GProp_GProps.hxx>
 #include <StlAPI_Writer.hxx>
 #include "TopoDS_Solid.hxx"
+#include <TopExp_Explorer.hxx>
 
 #include "EngrCAD.OCWrapper.h"
 
@@ -19,6 +21,8 @@
 #include <BRepAlgoAPI_Common.hxx>
 #include <BRepPrimAPI_MakeCylinder.hxx>
 #include <BRepPrimAPI_MakeCone.hxx>
+#include <BRepFilletAPI_MakeFillet.hxx>
+
 
 
 #pragma comment(lib, "TKernel.lib")
@@ -30,7 +34,7 @@
 #pragma comment(lib, "TKMesh.lib")
 #pragma comment(lib, "TKBO.lib")
 #pragma comment(lib, "TKG3d.lib")
-
+#pragma comment(lib, "TKOffset.lib")
 #pragma comment(lib, "TKPrim.lib")
 #pragma comment(lib, "TKV3d.lib")
 #pragma comment(lib, "TKIGES.lib")
@@ -39,6 +43,7 @@
 #pragma comment(lib, "TKStl.lib")
 #pragma comment(lib, "TKVrml.lib")
 #pragma comment(lib, "TKLCAF.lib")
+#pragma comment(lib, "TKFillet.lib")
 
 static TCollection_AsciiString toAsciiString(System::String^ theString)
 {
@@ -110,8 +115,7 @@ namespace EngrCADOCWrapper {
 
     NativeWrapper^ NativeWrapper::Translate(float x, float y, float z)
     {
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape shape = *shape_pointer;
+        TopoDS_Shape shape = *m_Impl;
 
         gp_Trsf transformation = gp_Trsf();
         transformation.SetTranslation(gp_Vec(x, y, z));
@@ -123,13 +127,40 @@ namespace EngrCADOCWrapper {
 
     }
 
+    System::Collections::Generic::List<FaceWrapper^>^ NativeWrapper::GetFaces()
+    {
+        TopoDS_Shape shape = *m_Impl;
+        int t = 0;
+        TopExp_Explorer Ex;
+        System::Collections::Generic::List<FaceWrapper^>^ list = gcnew System::Collections::Generic::List<FaceWrapper^>();
+        for (Ex.Init(shape, TopAbs_FACE); Ex.More(); Ex.Next()) {
+
+            TopoDS_Shape* t = new TopoDS_Shape(Ex.Current());
+            FaceWrapper^ face = gcnew FaceWrapper(t);
+            list->Add(face);
+        }
+        return list;
+    }
+
+    System::Collections::Generic::List<EdgeWrapper^>^ NativeWrapper::GetEdges()
+    {
+        TopoDS_Shape shape = *m_Impl;
+        int t = 0;
+        TopExp_Explorer Ex;
+        System::Collections::Generic::List<EdgeWrapper^>^ list = gcnew System::Collections::Generic::List<EdgeWrapper^>();
+        for (Ex.Init(shape, TopAbs_EDGE); Ex.More(); Ex.Next()) {
+
+            TopoDS_Edge* t = static_cast<TopoDS_Edge*>(new TopoDS_Shape(Ex.Current()));
+            EdgeWrapper^ face = gcnew EdgeWrapper(t);
+            list->Add(face);
+        }
+        return list;
+    }
+
     NativeWrapper^ NativeWrapper::Subtract(NativeWrapper^ other)
     {
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape difference = *shape_pointer;
-
-        TopoDS_Shape* other_pointer = static_cast<TopoDS_Shape*>(other->m_Impl);
-        TopoDS_Shape other_shape = *other_pointer;
+        TopoDS_Shape difference = *m_Impl;
+        TopoDS_Shape other_shape = *other->m_Impl;
 
         BRepAlgoAPI_Cut algo = BRepAlgoAPI_Cut(difference, other_shape);
         //differenceCut.SetFuzzyValue(0.1);
@@ -144,12 +175,8 @@ namespace EngrCADOCWrapper {
 
     NativeWrapper^ NativeWrapper::Union(NativeWrapper^ other)
     {
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape first = *shape_pointer;
-
-        TopoDS_Shape* other_pointer = static_cast<TopoDS_Shape*>(other->m_Impl);
-        TopoDS_Shape other_shape = *other_pointer;
-
+        TopoDS_Shape first = *m_Impl;
+        TopoDS_Shape other_shape = *other->m_Impl;
 
         BRepAlgoAPI_Fuse algo = BRepAlgoAPI_Fuse(first, other_shape);
         //combinedFuse.SetFuzzyValue(0.1);
@@ -162,11 +189,8 @@ namespace EngrCADOCWrapper {
 
     NativeWrapper^ NativeWrapper::Intersect(NativeWrapper^ other)
     {
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape first = *shape_pointer;
-
-        TopoDS_Shape* other_pointer = static_cast<TopoDS_Shape*>(other->m_Impl);
-        TopoDS_Shape other_shape = *other_pointer;
+        TopoDS_Shape first = *m_Impl;
+        TopoDS_Shape other_shape = *other->m_Impl;
 
 
         BRepAlgoAPI_Common algo = BRepAlgoAPI_Common(first, other_shape);
@@ -178,12 +202,41 @@ namespace EngrCADOCWrapper {
         return f;
     }
 
+    NativeWrapper^ NativeWrapper::Shell(double thickness)
+    {
+        TopoDS_Shape shape = *m_Impl;
+
+        TopTools_ListOfShape facesToRemove = TopTools_ListOfShape();
+        BRepOffsetAPI_MakeThickSolid builder = BRepOffsetAPI_MakeThickSolid();
+        builder.MakeThickSolidByJoin(shape, facesToRemove, thickness, 1e-4);
+
+
+        TopoDS_Shape* retVal = new TopoDS_Shape(builder.Shape());
+        NativeWrapper^ f = gcnew NativeWrapper(retVal);
+        return f;
+    }
+
+    NativeWrapper^ NativeWrapper::Round(System::Collections::Generic::List<RadiusDefinition^>^ definitions)
+    {
+        TopoDS_Shape shape = *m_Impl;
+        BRepFilletAPI_MakeFillet fillet = BRepFilletAPI_MakeFillet(shape);
+
+        for each (RadiusDefinition^ def in definitions) {
+            for each (EdgeWrapper^ edge in def->Edges) {
+                TopoDS_Edge* edge_pointer = edge->GetPointer();
+                fillet.Add(def->Radius, *edge_pointer);
+            }
+        }
+        TopoDS_Shape* retVal = new TopoDS_Shape(fillet.Shape());
+        NativeWrapper^ f = gcnew NativeWrapper(retVal);
+        return f;
+    }
+
     void NativeWrapper::SaveSTP(System::String^ path)
     {
         STEPControl_Writer writer;
 
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape shape = *shape_pointer;
+        TopoDS_Shape shape = *m_Impl;
 
         writer.Transfer(shape, STEPControl_AsIs);
         TCollection_AsciiString temp = toAsciiString(path);
@@ -194,8 +247,7 @@ namespace EngrCADOCWrapper {
     {
         StlAPI_Writer writer;
         writer.ASCIIMode() = true;
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape shape = *shape_pointer;
+        TopoDS_Shape shape = *m_Impl;
 
         BRepMesh_IncrementalMesh mesh(shape, 0.001f, true);
         mesh.Perform();
@@ -207,8 +259,7 @@ namespace EngrCADOCWrapper {
 
     float NativeWrapper::CalculateVolume()
     {
-        TopoDS_Shape* shape_pointer = static_cast<TopoDS_Shape*>(m_Impl);
-        TopoDS_Shape shape = *shape_pointer;
+        TopoDS_Shape shape = *m_Impl;
 
         GProp_GProps gprops;
 
