@@ -2,10 +2,68 @@
 
 
 namespace EngrCADOCWrapper {
+
+
+
+	public ref class FaceWrapper
+	{
+	public:
+		FaceWrapper(void* shape) : m_Impl(shape) {
+		}
+
+		~FaceWrapper() {
+			this->!FaceWrapper();
+		}
+	protected:
+		!FaceWrapper() {
+			delete m_Impl;
+		}
+
+	public:
+		void* GetPointer() { return m_Impl; }
+
+
+	private:
+		void* m_Impl;
+	};
+
+
+	public ref class EdgeWrapper
+	{
+	public:
+		EdgeWrapper(TopoDS_Edge* shape) : m_Impl(shape) {
+		}
+
+		~EdgeWrapper() {
+			this->!EdgeWrapper();
+		}
+	protected:
+		!EdgeWrapper() {
+			delete m_Impl;
+		}
+
+	public:
+		TopoDS_Edge* GetPointer() { return m_Impl; }
+
+
+	private:
+		TopoDS_Edge* m_Impl;
+	};
+
+	public ref class RadiusDefinition
+	{
+	public:
+		explicit RadiusDefinition(float radius, System::Collections::Generic::List<EdgeWrapper^>^ edges) : Radius(radius), Edges(edges)
+		{
+		}
+		System::Collections::Generic::List<EdgeWrapper^>^ Edges;
+		float Radius;
+	};
+
 	public ref class NativeWrapper
 	{
 	public:
-		NativeWrapper(void* shape) : m_Impl(shape) {
+		NativeWrapper(TopoDS_Shape* shape) : m_Impl(shape) {
 		}
 
 		~NativeWrapper() {
@@ -28,13 +86,16 @@ namespace EngrCADOCWrapper {
 
 		NativeWrapper^ Translate(float x, float y, float z);
 
+		System::Collections::Generic::List<FaceWrapper^>^ GetFaces();
+		System::Collections::Generic::List<EdgeWrapper^>^ GetEdges();
 
 		//Operations
 		NativeWrapper^ Subtract(NativeWrapper^ other);
 		NativeWrapper^ Union(NativeWrapper^ other);
 		NativeWrapper^ Intersect(NativeWrapper^ other);
-
-
+		NativeWrapper^ Shell(double thickness);
+		
+		NativeWrapper^ Round(System::Collections::Generic::List<RadiusDefinition^>^ definitions);
 
 		void SaveSTP(System::String^ path);
 		void SaveSTL(System::String^ path);
@@ -42,7 +103,9 @@ namespace EngrCADOCWrapper {
 		float CalculateVolume();
 
 	private:
-		void* m_Impl;
+		TopoDS_Shape* m_Impl;
 	};
+
+
 
 }
